+++ docs/release_notes.rst
@@ -9,21 +9,77 @@ Release Notes
 Upcoming Release
 ================
 
-Updated technologies
- - updated cost assumptions for 'digestible biomass to hydrogen' and "solid biomass to hydrogen"
-- Fix: Unit for methanation investment costs is now correctly displayed as "EUR/kW_CH4" (`#82 <https://github.com/PyPSA/technology-data/issues/82#event-8638160137>`_)
+* Add floating wind cost data
 
+* Add biomass-to-methanol route from DEA.
+
+* Add electric compression losses for hydrogen and gas pipelines from DEA.
+
+* Add methanol-to-kerosene from `Concawe report <https://www.concawe.eu/wp-content/uploads/Rpt_22-17.pdf>`_.
+
+* Add methanol-to-olefins/aromatics and steam cracker mostly from `DECHEMA report <https://dechema.de/dechema_media/Downloads/Positionspapiere/Technology_study_Low_carbon_energy_and_feedstock_for_the_European_chemical_industry.pdf>`_.
+
+* Added FOM for enhanced geothermal systems.
+
+Technology-Data 0.6.2 (7 August 2023)
+=====================================
+
+* Use DEA electrolysis and fuel cell assumptions by default.
+
+* Add steam generation of methanolisation process.
+
+* Use HVDC submarine cable cost from Härtel et al. (2017).
+
+Technology-Data 0.6.1 (4 August 2023)
+===========================================
 
 * New technologies
- - new biomass technologies ('biogas CC', 'central gas CHP CC', 'central hydrogen CHP', 'central solid biomass CHP CC', 'central solid biomass CHP powerboost CC',
+  - direct iron ore reduction (cost assumptions, conversion efficiencies)
+  - dry bulk carrier Capesize (cost assumptions)
+  - electric arc furnace (cost assumptions, conversion efficiencies)
+  - shipping fuel methanol (cost assumptions, emission intensity)
+  - iron ore DRI-ready (cost assumptions)
+
+Technology-Data 0.6.0 (24 May 2023)
+===========================================
+
+* General:
+  - Fix 'further_description' column from 'manual_inputs.csv' not being correctly parsed by the workflow
+  - Adjust electrolysis currency year to 2015
+
+* Updated technologies
+  - updated cost assumptions for 'digestible biomass to hydrogen' and "solid biomass to hydrogen"
+  - Fix: Unit for methanation investment costs is now correctly displayed as "EUR/kW_CH4" (`#82 <https://github.com/PyPSA/technology-data/issues/82#event-8638160137>`_)
+  - Fix source and description for 'solar' and 'solar-rooftop' to correctly indicate how they are calculated
+
+* New technologies
+  - 17 new energy storage technologies
+  - new biomass technologies ('biogas CC', 'central gas CHP CC', 'central hydrogen CHP', 'central solid biomass CHP CC', 'central solid biomass CHP powerboost CC',
 'direct firing gas', 'direct firing gas CC', 'direct firing solid fuels', 'direct firing solid fuels CC', 'electrobiofuels', 'solid biomass boiler steam CC', 'waste CHP', 'waste CHP CC',
 pelletizing cost for pellets from solid biomass residues)
+  - "utility-scale single-axis tracking" PV (cost assumptions)
+  - H2 liquefaction (Conversion efficiency)
+  - CH4 liquefaction (Conversion efficiency)
+  - CO2 liquefaction (Conversion efficiency)
+  - LOHC hydrogenation (Conversion efficiency)
+  - Ammonia crackier (Conversion efficiency)
+  - Steam methane reforming (Conversion efficiency)
+  - Methanol steam reforming (Conversion efficiency)
+  - Fischer-Tropsch (Conversion efficiency)
+  - seawater RO desalination (Conversion efficiency)
+  - Haber-Bosch (Conversion efficiency)
+  - air separation unit (Conversion efficiency)
+  - direct air capture (Conversion efficiency)
+  - Added data for Enhanced Geothermal Systems, as given by Aghahosseini, Breyer 2020
 
 * Changed technologies
- -
+  - methanation (Conversion efficiency)
+  - methanolisation (Conversion efficiency)
 
 * Features
  - energy penalties for biomass usages, biogas and gas boilers with carbon capture (calculations will be provided in an upcoming paper
+* Bug fixes
+  - Fixed a bug that ommited 'further description' from manually added data
 
 Technology-Data 0.5.0 (08 Februrary 2023)
 ===========================================
@@ -36,6 +92,7 @@ Technology-Data 0.5.0 (08 Februrary 2023)
   - new biomass technologies (BioSNG, BtL, biogas, biogas plus hydrogen, digestible biomass,digestible biomass to hydrogen, electric boiler steam, gas boiler steam, industrial heat pump high temperature, solid biomass boiler steam, solid bioass to hydrogen, biomass boiler for decentral heating
   - hydrogen storage tank type 1: Low pressure hydrogen tank storage (up to 200 bar)
   - hydrogen storage compressor: Compressor for filling hydrogen storage tanks (compression from 30 to 250 bar)
+  - 18 new energy storage technologies from PNNL "Energy Storage Grand Challenge Cost and Performance Assessment 2022"
 
 * Enable easy debugging of scripts by allowing python execution/ debugging in scripts